Move proof runner to pipeline package, add logger and CLI improvements
- Move proof runner from apps/api to packages/pipeline/src/proof.ts so
  the full pipeline (extract → storyboard → proof) can run from CLI
- Add lightweight logger utility (silent/error/info levels) to @adt/llm,
  replacing verbose boolean — CLI uses "silent", API uses default "info"
- Slim API proof-runner.ts to thin wrapper that manages API key env var
- Add proof steps (image-captioning, glossary, quiz-generation) to
  integration test with regenerated cache fixtures (12 LLM calls)
- Update CLI: run full pipeline end-to-end, progress bars for proof stage
- Export runProof and RunProofOptions from @adt/pipeline
