+++ packages/llm/src/index.ts
@@ -26,3 +26,5 @@ export {
 export { createPromptEngine, type PromptEngine } from "./prompt.js"
 
 export { createRateLimiter, type RateLimiter } from "./rate-limiter.js"
+
+export { createLogger, type LogLevel, type Logger } from "./logger.js"
