--- packages/pipeline/src/__tests__/fixtures/raven-cache/16209bb87344854a02a00e8ac38e18948336692edfe918ccd173680eb7f4c70e.json
@@ -1,15 +1,0 @@
-{
-  "reasoning": "This page contains a single story sentence paired with one illustration of Hyena and a bird, forming one coherent reading unit. The page number 2 is visible at the bottom center. Background is predominantly white with black text.",
-  "sections": [
-    {
-      "section_type": "text_and_single_image",
-      "part_ids": [
-        "pg003_gp001",
-        "pg003_im002"
-      ],
-      "background_color": "#ffffff",
-      "text_color": "#000000",
-      "page_number": 2
-    }
-  ]
-}
