--- packages/pipeline/src/__tests__/fixtures/raven-cache/16fb9f81b6c88b916b7631297f5e497a0ec09b68c3758227f07c57e7ce7c865c.json
@@ -1,11 +1,0 @@
-{
-  "title": "Hyena and Raven",
-  "authors": [
-    "Tony Lelliott",
-    "Wiehan de Jager"
-  ],
-  "publisher": "African Storybook",
-  "language_code": "en",
-  "cover_page_number": 1,
-  "reasoning": "Page 1 appears to be the front cover because it contains the book title (“Hyena and Raven”), the creator names (“Tony Lelliott”, “Wiehan de Jager”), a prominent illustrative design, and the African Storybook branding/logo. No earlier blank page is provided. The narrative text on pages 2–3 is in English (e.g., “Hyena and Raven were once great friends.”), so the language is English (en). The publisher is inferred from the visible “African Storybook.org” mark on the cover."
-}
