+++ packages/pipeline/src/__tests__/fixtures/raven-cache/1ecb4f6c0c7446bdfd65f158037000ca71c5d16c4758f2b8ec629b692b462921.json
@@ -1,0 +1,9 @@
+{
+  "captions": [
+    {
+      "image_id": "pg003_im002",
+      "reasoning": "The image shows a close-up, cartoon-style scene: a hyena faces a black bird (looks like a crow) against a blue sky with a few white clouds. The hyena’s paw is extended toward the bird as if talking or making a request, matching the story text about wanting to fly.",
+      "caption": "A hyena leans toward a black bird and reaches out a paw as they face each other under white clouds in a blue sky."
+    }
+  ]
+}
