+++ packages/pipeline/src/__tests__/fixtures/raven-cache/244cdaa25b94c31f5e149b91466382e4aa5a23edb399256966514d604a3f612a.json
@@ -1,0 +1,11 @@
+{
+  "title": "Hyena and Raven",
+  "authors": [
+    "Tony Lelliott",
+    "Wiehan de Jager"
+  ],
+  "publisher": "African Storybook.org",
+  "language_code": "en",
+  "cover_page_number": 1,
+  "reasoning": "Page 1 has the full title ('Hyena and Raven'), creator names ('Tony Lelliott', 'Wiehan de Jager'), and a publisher/brand mark ('African Storybook.org') in a layout typical of a front cover. Pages 2–3 contain the story text with page numbers, indicating they are interior pages. The narrative text on pages 2–3 is in English, so the language code is 'en'. The publisher name is taken from the African Storybook.org logo on the cover."
+}
