+++ packages/pipeline/src/__tests__/fixtures/raven-cache/350c01c3a59cf75cce16bcfe653d7bffcc0ecd9aa7df72befc90dbf5c149091f.json
@@ -1,0 +1,9 @@
+{
+  "captions": [
+    {
+      "image_id": "pg002_im002",
+      "reasoning": "The image shows two main characters from the story: a spotted hyena standing on the ground and a black raven flying above it. The caption should describe their positions and the friendly, playful mood implied by the scene.",
+      "caption": "A spotted hyena grins on the dusty ground while a black raven flaps its wings overhead against a green sky."
+    }
+  ]
+}
