+++ packages/pipeline/src/__tests__/fixtures/raven-cache/581c2311562dd1a6888169fad266bec8e565bbb2bc1bccf3a8c48310027dd75e.json
@@ -1,5 +1,5 @@
 {
-  "reasoning": "The page contains two short narrative sentences set as separate text blocks on the right side, and a single page number centered at the bottom. No headers, titles, captions, or other text appear on the illustration.",
+  "reasoning": "The page contains two short narrative paragraphs on the right side and a centered page number at the bottom. No headers, captions, or other text appears within the illustration.",
   "groups": [
     {
       "group_type": "paragraph",
