--- packages/pipeline/src/__tests__/fixtures/raven-cache/92eec6dae93fbf9b48c467a1f7ff1873738d4b04a3e00a49ec369a9de70804e6.json
@@ -1,18 +1,0 @@
-{
-  "reasoning": "This page is a front cover: it has the story title, authors/illustrators, language label, publisher/website mark, and a full-page cover illustration. All elements belong together as one coherent cover unit.",
-  "sections": [
-    {
-      "section_type": "front_cover",
-      "part_ids": [
-        "pg001_im001",
-        "pg001_im002",
-        "pg001_gp001",
-        "pg001_gp002",
-        "pg001_gp003"
-      ],
-      "background_color": "#e89010",
-      "text_color": "#000000",
-      "page_number": null
-    }
-  ]
-}
