+++ packages/pipeline/src/__tests__/fixtures/raven-cache/b9ccb98202b3c9fe51028632cf478c88893688aecaaa8d33fc22519c1025be32.json
@@ -1,5 +1,5 @@
 {
-  "reasoning": "The page contains one block of narrative text at the upper right and a centered page number at the bottom. No additional captions or labels appear within the illustration.",
+  "reasoning": "The page contains one narrative sentence split across multiple lines on the right side, plus a page number at the bottom center.",
   "groups": [
     {
       "group_type": "paragraph",
