--- packages/pipeline/src/__tests__/fixtures/raven-cache/d2ca2c4b742df8ba33f239228fce87399248059c4918380d5f56bf5fe232e93d.json
@@ -1,16 +1,0 @@
-{
-  "reasoning": "This page is a single narrative moment introducing the characters and their friendship, supported by one illustration. The two paragraphs belong together with the image as one coherent reading section.",
-  "sections": [
-    {
-      "section_type": "text_and_single_image",
-      "part_ids": [
-        "pg002_gp001",
-        "pg002_gp002",
-        "pg002_im002"
-      ],
-      "background_color": "#ffffff",
-      "text_color": "#000000",
-      "page_number": 1
-    }
-  ]
-}
