+++ packages/pipeline/src/__tests__/fixtures/raven-cache/d35f7cc9b7f08ff0d127b3eb941762c493facff8cba943facb89a60aaa7c6a82.json
@@ -1,0 +1,18 @@
+{
+  "reasoning": "This page is a front cover: it contains the story title, author/illustrator credits, language label, publisher/site mark, and a single main illustration that serves as cover art. All elements belong together as one cover unit; no page number is visible.",
+  "sections": [
+    {
+      "section_type": "front_cover",
+      "part_ids": [
+        "pg001_im001",
+        "pg001_im002",
+        "pg001_gp001",
+        "pg001_gp002",
+        "pg001_gp003"
+      ],
+      "background_color": "#e58c1f",
+      "text_color": "#000000",
+      "page_number": null
+    }
+  ]
+}
