+++ packages/pipeline/src/__tests__/fixtures/raven-cache/dd173b5f5b06d43d2b29ca572322bb9e093c60e22c0814361c77516bbdd6bc78.json
@@ -1,0 +1,15 @@
+{
+  "reasoning": "This page contains a single narrative sentence paired with one illustration showing Hyena speaking to a bird. They form one coherent reading unit. The visible page number is 2 at the bottom center. Background is predominantly white with black text.",
+  "sections": [
+    {
+      "section_type": "text_and_single_image",
+      "part_ids": [
+        "pg003_gp001",
+        "pg003_im002"
+      ],
+      "background_color": "#ffffff",
+      "text_color": "#000000",
+      "page_number": 2
+    }
+  ]
+}
