+++ packages/pipeline/src/pipeline.ts
@@ -3,7 +3,7 @@ import path from "node:path"
 import { createBookStorage } from "@adt/storage"
 import type { Storage } from "@adt/storage"
 import { createLLMModel, createPromptEngine, createRateLimiter } from "@adt/llm"
-import type { LLMModel } from "@adt/llm"
+import type { LLMModel, LogLevel } from "@adt/llm"
 import { extractPDF } from "./pdf-extraction.js"
 import { extractMetadata, buildMetadataConfig } from "./metadata-extraction.js"
 import { classifyPageText, buildClassifyConfig } from "./text-classification.js"
@@ -36,6 +36,8 @@ export interface RunPipelineOptions {
   templatesDir: string
   /** Override cache directory. Defaults to {booksRoot}/{label}/.cache */
   cacheDir?: string
+  /** LLM console log level. Defaults to "info". Use "silent" for no output. */
+  logLevel?: LogLevel
 }
 
 export async function runPipeline(
@@ -52,6 +54,7 @@ export async function runPipeline(
     configPath,
     promptsDir,
     templatesDir,
+    logLevel,
   } = options
 
   if (!fs.existsSync(pdfPath)) {
@@ -85,6 +88,7 @@ export async function runPipeline(
       cacheDir,
       promptEngine,
       rateLimiter,
+      logLevel,
       onLog: (entry) => storage.appendLlmLog(entry),
     })
 
@@ -126,6 +130,7 @@ export async function runPipeline(
           cacheDir,
           promptEngine,
           rateLimiter,
+          logLevel,
           onLog: (entry) => storage.appendLlmLog(entry),
         })
       : null
@@ -140,6 +145,7 @@ export async function runPipeline(
       cacheDir,
       promptEngine,
       rateLimiter,
+      logLevel,
       onLog: (entry) => storage.appendLlmLog(entry),
     })
     const renderModels = new Map<string, LLMModel>()
@@ -151,6 +157,7 @@ export async function runPipeline(
         cacheDir,
         promptEngine,
         rateLimiter,
+        logLevel,
         onLog: (entry) => storage.appendLlmLog(entry),
       })
       renderModels.set(modelId, model)
